+++ poc/classes/AuxSTClassInstance.py
@@ -1,12 +1,11 @@
-from poc.classes.AuxST import AuxSTBlock
+from poc.classes.AuxSTInstance import AuxSTInstance
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
 
-class AuxSTClassInstance(AuxSTBlock):
+class AuxSTClassInstance(AuxSTInstance):
 
     def __init__(self, i):
-        super().__init__(AuxSymbolTable.property, i)
-        self.def_type = AuxSymbolTable.classProperty
+        super().__init__(i)
+        self.def_type = AuxSymbolTable.classInstance
         self.zfrom = i.corrected_position('VariableType')
         self.zto = i.last_positions_by_rule['ClassInstance'].pos_to_str()
-        self.mandatory = False