+++ poc/classes/AuxSTDefinitionFunctionalTerm.py
@@ -1,8 +1,8 @@
-from poc.classes.AuxST import AuxSTBlock
+from poc.classes.AuxSTBlockWithSignature import AuxSTBlockWithSignature
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
 
-class AuxSTDefinitionFunctionalTerm(AuxSTBlock):
+class AuxSTDefinitionFunctionalTerm(AuxSTBlockWithSignature):
 
     def __init__(self, i):
         super().__init__(AuxSymbolTable.block_def, i)